+++ Client/Scripts/Gameplay/Interfaces/IPingMeasurer.cs
@@ -2,5 +2,5 @@ namespace Soteo.Gameplay.Interfaces;
 
 public interface IPingMeasurer
 {
-    float? Ping(Guid peerId);
+    double? Ping(Guid peerId);
 }
